fix: Use iter_batched() and iter_batched_ref() for the remaining benchmarks
The bulk of the changes was done Claude Sonnet 4. Additionally I moved
`DVector` allocations outside of the benchmark, and added anything
allocated and not consumed into a return tuple of a benchmark closure to
ensure that implicit drop/free is not included into the measured time.

This fixes https://github.com/dimforge/nalgebra/issues/1547 for the
remaining benchmarks.